Fixing some unit tests
I realized that I'm stupid and the BasicPlayer will need to select cards
based on the cards returned by get_valid_plays(). This means that, we
need to provide (in some way) a reduced set of cards to each of these
functions, or provide the entire game_state. I've chosen the former, to
make testing easier, and the fact that these functions should be able to
be called with or without this reduced list.