Reworking unit tests some more.
It became difficult to just pass around the game state object.
Additionally, it dawned on me that I needed to be sure that the cards
chosen were legal cards being played. Why look for trump cards, if I
can't play one? Therefore we have now changed many of the functions to
take a list of cards that can be played. This list can be generated from
get_valid_plays() or it can be self.hand. Either will work in the
correct situation.